Initial who talks routing
The tally will now present users with a decision: to count time or people.
The who talks form has buttons to track time, and then the same standard
title and hashtag fields for use in the tally generation.

Right now a placeholder talkschart endpoint has been set up, which will
be populated with similar logic to the previous chart endpoint.