Headcounts use for who talks
The who talks workflow has been updated to now collect information about
who is involved in a conversation.  This allows the app to generate more
accurate descriptions of a situation, and also judge whether or not women
were actually speaking a fair amount of the time based on the number of
women in the room.