Add nonbinary and whotalks icons
Upfront: sorry for combining two features in one commit, won't happen
again I swear.

This commit contains two improvements.

First, it adds support for counting nonbinary persons to the core
tally.  This meant the form became long enough that we felt it
would be best to split it into two screens.  If there are no
nonbinary persons we do not include the count on the resulting
tally, but when they are there they are counted.  Hooray :D

Second, it adds icons and description text to the who talks charts.
The who talks charts were hard to parse and understand as being
good or bad without some guidance.  There was a lot of discussion
about this and it went both ways, but we tried it out without them
and got the feedback so here we are.

We're going with a slightly more harsh 10% variance for men talking
before breaking out the BLAH BLAH BLAH.  This is only noteworthy
because in the past we used a 20%-from-ideal heuristic.  This
would mean if men made up 50% of a panel and spoke more than 55%
of the time (5% is 10% of 50%) then we would show a blah blah blah.

Issue #7 Who talks judgment accuracy
Issue #19 Support counting of non-binary persons