Change unlink to unlinksync
At some point in the past four years Node changed unlink to require
a callback. This commit replaces unlink with the thread blocking
unlinkSync.

In reality there is no reason to block, but it would be worth doing
a more general optimization review and refactor, and sync seems less
likely to result in weird race-condition related bugs.

Issue #24 Update to Node 10.5.x