Improve pause button UI
In the Who Talks timing screen the user has always been able to pause
the current timer by pressing the button a second time.  This is not
obvious in any way from the current UX.  This commit swaps out the text
on the active button so it says "Pause."

That change meant that it was no longer easy to know who was talking at
a glance, so the button also slightly modifies the "Who's Talking" text.

The result still has some room for improvement, but it seems to be
better than what was there before.

Issue #21 Who's Talking pause button needed