Fix 40% edge case
The logic for deciding how good a ratio was for whotalks and also for
the normal tally had a mistake in that it always compared exactly > or
exactly <.  This meant the "exactly 40%" case was not captured and the
resulting category was wrong.

This makes it so >= 40% will be counted as sunny and bright.