Add back buttons to tally flow
Users in the browser could use back buttons, but this didn't really help
for mobile users.  Now back buttons are part of the interface.

Because of the way that Who Talks works, we can't add back buttons there
just yet (we need to do a better job of storing who talks results and
re-rendering those results on refresh first).

Issue #45 Add a back button