Update all dependencies
The dependencies were fundamentally broken in node 11, so this is
a baseline reset for all of them to the "latest" based on yarn add.

This commit also removes new relic functionality (which was spitting out
errors anyway).  This is because new relic does not support node 11.

It also replaces node-uuid with uuid, since node-uuid was deprecated and
the maintainer has renamed the package to uuid.

Issue #55