Firebase and imgur modernization
Both firebase and imgur updates had breaking changes which had to be
addressed.  For imgur it was just a matter of changing the name of the
method we call to upload a file, but for firebase they have
fundamentally changed the way authentication works and have depricated
the use of secrets (opting for admin credentials instead).

Issue #55