Add back button to first tally page
If you decide to count folks you should be able to go back to pick
a different type of tally without needing to know that you can tap the
logo.  This adds a back button to the first page of the tally.  It also
adjusts some of the layout features of the page.

Issue #45