Support versions on server
We need to make it possible for different versions of the app to be able
to have different content served.  For example, old versions of the app
should not have the download button appear since the app is not able to
process that.

Eventually we may want multiple deployed instances, but for now this
version cookie should allow us to conditionally render the download
feature.  Eventually the conditional will be removed.