Add who talks back button
The who talks back button was more complex because of the way who talks
sets the timers.  This commit adds that button in, though we should put
some urgency on the ability to edit the timers manually in order to
cover the use case where a user actually WANTS to start from a clean
slate.

Issue #45