Fix the fixed footer image
The footer image would fail to stick to the bottom of the page in the
event of a scroll.  This fixes that up, making a guaranteed space for
the footer as well.

This is in praparatin for BallotMANia which is expected to have longer
forms than we have seen so far.

This also fixes a few tiny bugs that I noticed while adding the footer.