Add photo search button
When a candidate has no photo we were showing the bio, but that rarely
solved the problem of wanting to know more about the person's racial
identity.  We're going to try linking to a google image search for the
person in hopes that it will address that use case.

Issue #65