Remove position from search
When there is no image we link to a google image search to help the user
identify the candidate.  We had been including the name of the position
they are running for, which in some cases caused google to get confused
(e.g. if that position is in a popular news headline unrelated to the
candidate).