Add custom social share
We want the ability to have custom social share defaults depending on
the type of tally and the outcome.  This adds a new attribute to tallies
which will create that concept.

This addition begins to highlight the need to refactor the tally; as we
have added new tally types we have created a fair bit of duplicate code.
It's manegable at the moment but does mean technical debt.

Issue #72