Allow non-selections
In Ballotmania there are situations where a user might not want to make
a selection.  In the past we did not allow them to submit, but after
feedback some folks are uncomfortable with that requirement.

This changes the logic so that users are alerted if they forget to make
a selection, but are ultimately allowed to submit if they choose to.

In the event of a non-selection the person is not counted in any way.

Issue #79