Remove race from ballotmania
After a few months of tests the biggest issue we faced was that folks
without photos were very difficult to identify in terms of race.  We
experimented with allowing google image search (which was noisy and
didn't solve the problem).

This removes the "Woman of color" buttons from ballotmania so that the
only thing counted is gender identity.

Issue #80