Allow manual whotalks entry
WhoTalks requires real time tracking of who is talking, but there are
times when that number needs to either be manually edited due to
a mistake in tracking OR the time totals were pre-calculated.

This changes the time display to an editable field, which allows people
to make either type of adjustment.

Issue #49