Add interface choice
Internal feedback indicated that the manual entry use case wasn't
clearly covered by just adding an "edit" button, and we need to be more
explicit.

This adds in a user decision screen before rendering the interface.
Depending on the user choice, the interface shown FIRST will either be
the timers or the manual edit mode.  In either case it is possible to
switch back and forth between the two interfaces in a single session
without losing data.

Issue #49