Set up a TypeScript project
This is a huge commit and it moves us into the world of TypeScript.  It
sets up a whole slew of baseline tooling for the repository, such as:

1. TypeScript itself
2. ESLint (for making sure the code looks nice)
3. Jest (for testing)

This also sets up some basic github workflows to invoke the lint and
tests on PR, and incorporates CodeCov for test coverage reports.

I did not delete any old code; I'd like to try to do that as I migrate
it to TypeScript.  That said, this repository no longer builds against
that code, so it really only exists as a temporary artefact to track
what has / hasn't been migrated.

Issue #115 Initialize the new repository