Set up SvelteKit
Svelte is a modern and straightforward frontend framework; SvelteKit is
somewhat more cutting edge but robust enough that I think it is worth
using.

This moves our repository to use the out-of-the-box setup for Svelte.
We will use a separate repository for any backend functionality.

This was built using:
```
npm create svelte@latest tally
```

For more information check out Svelte[1]

[1] https://kit.svelte.dev/

Issue #117