Create landing page
This is a somewhat larger commit since in order to support the landing
page we also needed to break ground on a few concepts including layouts,
theming, and a few components.

The current theme approach is powered by a json file that defines some
css variables and content strings; this results in a small flash of
unstyled content which will be an issue to fix in a future issue.

Issue #117 Set up a landing page