Create themed chart
This applies the FQ theme to the project and also renders a first
version of the chart.

Ultimately this commit is merging a large number of things at once; this
is due to the nature of development where I switched gears from tally
development into FQ development mid-stream.  I will do better in future.
This commit does the following specifically:

1. Logic for rendering a chart
2. Removing elements that aren't used by SEA
3. Updating theme colors and logos to SEA

Some of this will ultimately be upstreamed back to the tally.