Stop collecting totals
Due to the way the tool evolved originally the interfaces were
collecting "totals" for each gender (e.g. "total women") and then
collecting women of color as a subset.

We think it will be more clear to collect just the subsets directly:
white women, and women of color (same for other genders).