Set up firebase
This creates some configuration and utility libraries for firebase which
can be used by our application to read / write against a cloud
firestore.

I'm including two varieties of the library for now.  This is because
I actually integrated with firebase in two different ways:

1. Using the firebase admin SDK, which can only run on the server side
   and provides full read / write access.
2. Using the firebase client SDK, which would require use of
   authorization and would allow us to avoid a server side need.

I had originally thought that the client side option wouldn't be
possible since we don't have users, but then recently learned that
Firebase has anonymous authentication as a viable mechanism for use.